refactor(quiz): move useQuizPhase hook to hooks directory and update ChoicesSection import
- Relocated the `useQuizPhase` hook to the hooks directory for better organization.
- Updated `ChoicesSection` to import `useQuizPhase` from the new location, enhancing modularity and clarity.
