+++ features/answer-quiz/components/ChoicesSection.tsx
@@ -2,7 +2,7 @@ import { ThemedText } from '@/components/ThemedText'
 import { PrimaryButton } from '@/components/ui/button/PrimaryButton'
 import { Tables } from '@/database.types'
 import { QuizChoice } from '@/features/answer-quiz/components/QuizChoice'
-import { useQuizPhase } from '@/features/answer-quiz/context/QuizPhaseContext'
+import { useQuizPhase } from '@/features/answer-quiz/hooks/useQuizPhase'
 import { useQuizAnswer } from '@/features/answer-quiz/hooks/useQuizAnswer'
 import { useQuizNavigation } from '@/features/answer-quiz/hooks/useQuizNavigation'
 import { useQuizChoices } from '@/features/answer-quiz/hooks/useQuizQuery'
